+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.1.1 (10/08/2014)
 
 - Add externs file for StripeCheckout.
+- Create `stripe.test` namespace.
 
 ## 0.1.0 (10/08/2014)